backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database  (SQLAlchemy 2.0.36+ fixes Python 3.14 FastIntFlag compat)
sqlalchemy>=2.0.36
alembic>=1.14.0
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
pgvector>=0.3.6

# Pydantic / Settings
pydantic[email]>=2.9.0
pydantic-settings>=2.5.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.12

# HTTP client
httpx>=0.28.0

# OpenAI
openai>=2.0.0

# ML / Data Science  (Python 3.14 compatible — no pinned build-time versions)
scikit-learn>=1.5.0
pandas>=2.2.2
numpy>=2.0.0
scipy>=1.14.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
aiosqlite>=0.20.0

# Utilities
aiofiles>=24.1.0
python-dotenv>=1.0.1
tenacity>=9.0.0
structlog>=24.4.0
